move operations to its own module
